// tests/data/enums.ts
export enum Tags {
  SMOKE = '@smoke',
  FLOW = '@flow',
  NEGATIVE = '@negative',
  CREATE = '@create',
  ROGUE = '@rogue',
  KAEL = '@kael',
  GET = '@get',
  POST = '@post',
  PATCH = '@patch',
  PUT = '@put',
  DELETE = '@delete',
}

export enum CharacterClass {
  ROGUE = 9,
}

export enum CharacterStatus {
  DRAFT = 'draft',
  IN_PROGRESS = 'in_progress',
  COMPLETE = 'complete',
}